Change `std_Group.project` type to closure.
The `project` field of the `std_Group` type maps `DDValue` to the inner
type of the group.  It was originally defined as a function.  Here we
generalize it to a closure.  The concrete use case we want to enable right
now came up when injecting debugging hooks in the program: given a group
over a two-tuple, we want to transform it into a group over the second
element of the tuple by simply changing the projection function.