Generate debug aggregate funcs per Aggregate operator.
Generating a debug aggregate per original aggregation function is not
sufficient when the function is implemented in Rust and uses trait
bounds that are not visible to DDlog.  This causes Rust compilation
errors since generic arguments to the generated function do not have the
right traits.  The solution is to generate a debug aggregate for each
call site specialized to concrete types at this call site.